Separate widgets into different files one per each widget.
