+++ src/app/root-content/dashboard/widgets/documents/documents-widget.component.html
@@ -1,0 +1,12 @@
+<div class="content widget-header">Documents
+	<div class="right floated">
+		<i class="refresh small icon widget-button" (click)="refreshDocumentsCount()"></i>
+		<i class="remove small icon widget-button" (click)="close($event, 'Documents')"></i>
+	</div>
+</div>
+<div class="content widget-body">
+	<div class="widget-text">{{documentsTotalCount}}</div>
+	<img class="widget-icon" src="/assets/images/documentIcon.png" height="60"/>
+</div>
+
+
