+++ src/app/root-content/dashboard/widgets/triples/triples-widget.component.html
@@ -1,0 +1,12 @@
+<div class="content widget-header">Triples
+	<div class="right floated">
+		<i class="refresh small icon widget-button" (click)="refreshTriplesCount()"></i>
+		<i class="remove small icon widget-button" (click)="close($event, 'Triples')"></i>
+	</div>
+</div>
+<div class="content widget-body">
+	<div class="widget-text">{{triplesTotalCount}}</div>
+	<img class="widget-icon" src="/assets/images/triplesIcon.png" height="60"/>
+</div>
+
+
