+++ src/app/root-content/dashboard/widgets/widget.component.ts
@@ -1,8 +1,4 @@
-import { Component, ElementRef } from "@angular/core";
-
-import { Class as Carbon } from "carbonldp/Carbon";
-
-import { WidgetsService } from "./widgets.service";
+import { Component } from "@angular/core";
 
 import "semantic-ui/semantic";
 
@@ -13,49 +9,5 @@ import "semantic-ui/semantic";
 })
 
 export class WidgetComponent{
-	carbon:Carbon;
-	element:ElementRef;
-	widgetsService:WidgetsService;
-	documentsTotalCount;
-	triplesTotalCount;
-
-
-	constructor( element:ElementRef, carbon:Carbon, widgetsService:WidgetsService ) {
-		this.element = element;
-		this.widgetsService = widgetsService;
-		this.carbon = carbon;
-	}
-
-	ngOnInit():void {
-		this.widgetsService.getDocumentsTotalCount().then( ( count ) => { this.documentsTotalCount = count; });
-		this.widgetsService.getTriplesTotalCount().then( ( count ) => { this.triplesTotalCount = count; });
-	}
-	
-	public refreshDocumentsCount(){
-		this.widgetsService.getDocumentsTotalCount().then( ( count ) => { this.documentsTotalCount = count; });
-	}
-
-	public refreshTriplesCount(){
-
-		this.widgetsService.getTriplesTotalCount().then( ( count ) => { this.triplesTotalCount = count; });
-	}
-
-	public close(e, widgetName){
-		let widget:Element;
-		let widgetsMenuItem:Element;
-		if( widgetName === "Triples") {
-			widget = document.querySelector(".widget-container--totalTriples");
-			widgetsMenuItem = document.querySelector(".widgetsMenu-item--totalTriples");
-		} else if ( widgetName === "Documents" ){
-			widget = document.querySelector(".widget-container--totalDocuments");
-			widgetsMenuItem = document.querySelector(".widgetsMenu-item--totalDocuments");
-		} else {
-			return false;
-		}
-
-		this.widgetsService.closeWidget(widget, widgetsMenuItem);
-
-	}
-
 
 }
