+++ src/app/root-content/dashboard/widgets/widgets-menu/widgets-menu.component.html
